src/transcription-service/requirements.txt
pyannote.audio
pyannote.core
scipy
torch
numpy
python-dotenv
openai-whisper
pydub
pymongo